+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - BaseColorPicker rendering so that it doesn't get cut off
 - definition of custom html-elements in Storybook (added try-catch statement)
 - joined JS-code of BaseCheckbox, BaseRadio and BaseSwitch
+- joined JS-code of BaseButton and BaseIconButton
 
 ## [0.14.0]